ratdispenser/requirements.txt
file_magic == 0.4.0
